ai/flask/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
flask>=1.0.0
flask_cors>=1.0.0